Split tests for eQ (URL gen.) and handlers (views)
