+++ tests/unit/test_eq.py
@@ -1,807 +1,16 @@
 import functools
-import json
-import time
-import uuid
 from unittest import mock
-from urllib.parse import urlsplit, parse_qs
 
-from aiohttp.client_exceptions import ClientConnectionError, ClientConnectorError
-from aiohttp.test_utils import AioHTTPTestCase, unittest_run_loop
+from aiohttp.test_utils import unittest_run_loop
 from aioresponses import aioresponses
 
-from app import (
-    BAD_CODE_MSG, BAD_RESPONSE_MSG, CODE_USED_MSG, INVALID_CODE_MSG, NOT_AUTHORIZED_MSG,
-    CONNECTION_ERROR_MSG, REDIRECT_FAILED_MSG, SERVER_ERROR_MSG)
-from app.app import create_app
 from app.eq import format_date, find_event_date_by_tag
-from app.exceptions import InvalidEqPayLoad, InactiveCaseError
-from app.handlers import join_iac, validate_case
+from app.exceptions import InvalidEqPayLoad
 
+from . import RHTestCase
 
-def skip_build_eq(func, *args, **kwargs):
-    """
-    Helper decorator for manually patching the methods of app.eq.EqPayloadConstructor.
 
-    This can be useful for tests that perform as a client but wish the server to skip the builder functionality.
-
-    The test case checks for and calls when possible .setUp and .tearDown attributes on each test method
-    at server setUp (setUpAsync) and server tearDown (tearDownAsync).
-
-    :param func: test method that requires the patch
-    :param args: the test method's arguments
-    :param args: the test method's keyword arguments
-    :return: new method with patching functions attached as attributes
-    """
-
-    async def _override_eq_payload_constructor(test_case, *_):
-        from app import eq
-
-        async def build(_):
-            return test_case.eq_payload
-
-        eq.EqPayloadConstructor._bk__init__ = eq.EqPayloadConstructor.__init__
-        eq.EqPayloadConstructor.__init__ = lambda *args: None
-        eq.EqPayloadConstructor._bk_build = eq.EqPayloadConstructor.build
-        eq.EqPayloadConstructor.build = build
-
-    async def _reset_eq_payload_constructor(*_):
-        from app import eq
-
-        eq.EqPayloadConstructor.__init__ = eq.EqPayloadConstructor._bk__init__
-        eq.EqPayloadConstructor.build = eq.EqPayloadConstructor._bk_build
-
-    @functools.wraps(func, *args, **kwargs)
-    def new_func(self, *inner_args, **inner_kwargs):
-        return func(self, *inner_args, **inner_kwargs)
-
-    new_func.setUp = _override_eq_payload_constructor
-    new_func.tearDown = _reset_eq_payload_constructor
-
-    return new_func
-
-
-def build_eq_raises(func, *args, **kwargs):
-    """
-    Helper decorator for manually patching the methods of app.eq.EqPayloadConstructor.
-
-    This can be useful for tests that perform as a client but wish the server to raise InvalidEqPayLoad when .build()
-    is called on an instance of app.eq.EqPayloadConstructor.
-
-    The test case checks for and calls when possible .setUp and .tearDown attributes on each test method
-    at server setUp (setUpAsync) and server tearDown (tearDownAsync).
-
-    :param func: test method that requires the patch
-    :param args: the test method's arguments
-    :param args: the test method's keyword arguments
-    :return: new method with patching functions attached as attributes
-    """
-
-    async def _override_eq_build_with_error(*_):
-        from app import eq
-
-        async def build(_):
-            raise InvalidEqPayLoad('')
-
-        eq.EqPayloadConstructor._bk__init__ = eq.EqPayloadConstructor.__init__
-        eq.EqPayloadConstructor.__init__ = lambda *args: None
-        eq.EqPayloadConstructor._bk_build = eq.EqPayloadConstructor.build
-        eq.EqPayloadConstructor.build = build
-
-    async def _reset_eq_payload_constructor(*_):
-        from app import eq
-
-        eq.EqPayloadConstructor.__init__ = eq.EqPayloadConstructor._bk__init__
-        eq.EqPayloadConstructor.build = eq.EqPayloadConstructor._bk_build
-
-    @functools.wraps(func, *args, **kwargs)
-    def new_func(self, *inner_args, **inner_kwargs):
-        return func(self, *inner_args, **inner_kwargs)
-
-    new_func.setUp = _override_eq_build_with_error
-    new_func.tearDown = _reset_eq_payload_constructor
-
-    return new_func
-
-
-def skip_encrypt(func, *args, **kwargs):
-    """
-    Helper decorator for manually patching the encrypt function in handlers.py.
-
-    This can be useful for tests that perform as a client but wish the server to skip encrypting a payload.
-
-    The test case checks for and calls when possible .setUp and .tearDown attributes on each test method
-    at server setUp (setUpAsync) and server tearDown (tearDownAsync).
-
-    :param func: test method that requires the patch
-    :param args: the test method's arguments
-    :param args: the test method's keyword arguments
-    :return: new method with patching functions attached as attributes
-    """
-
-    async def _override_sdc_encrypt(*_):
-        from app import handlers
-
-        def encrypt(payload, **_):
-            return json.dumps(payload)
-
-        handlers._bk_encrypt = handlers.encrypt
-        handlers.encrypt = encrypt
-
-    async def _reset_sdc_encrypt(*_):
-        from app import handlers
-
-        handlers.encrypt = handlers._bk_encrypt
-
-    @functools.wraps(func, *args, **kwargs)
-    def new_func(self, *inner_args, **inner_kwargs):
-        return func(self, *inner_args, **inner_kwargs)
-
-    new_func.setUp = _override_sdc_encrypt
-    new_func.tearDown = _reset_sdc_encrypt
-
-    return new_func
-
-
-class TestGenerateEqURL(AioHTTPTestCase):
-
-    region_code = 'GB-ENG'
-    language_code = 'en'
-
-    start_date = '2018-04-10'
-    end_date = '2020-05-31'
-    return_by = '2018-05-08'
-
-    async def get_application(self):
-        return create_app('TestingConfig')
-
-    async def setUpAsync(self):
-        test_method = getattr(self, self._testMethodName)
-        if hasattr(test_method, 'setUp'):
-            await test_method.setUp(self)
-
-    async def tearDownAsync(self):
-        test_method = getattr(self, self._testMethodName)
-        if hasattr(test_method, 'tearDown'):
-            await test_method.tearDown(self)
-
-    def assertLogLine(self, watcher, event, **kwargs):
-        """
-        Helper method for asserting the contents of structlog records caught by self.assertLogs.
-
-        Fails if no match is found. A match is based on the main log message (event) and all additional
-        items passed in kwargs.
-
-        :param watcher: context manager returned by `with self.assertLogs(LOGGER, LEVEL)`
-        :param event: event logged; use empty string to ignore or no message
-        :param kwargs: other structlog key value pairs to assert for
-        """
-        for record in watcher.records:
-            message_json = json.loads(record.message)
-            try:
-                if (
-                    event in message_json.get('event', '')
-                    and all(message_json[key] == val for key, val in kwargs.items())
-                ):
-                    break
-            except KeyError:
-                pass
-        else:
-            self.fail(f'No matching log records present: {event}')
-
-    def setUp(self):
-        super().setUp()  # NB: setUp the server first so we can use self.app
-        with open('tests/test_data/case/case.json') as fp:
-            self.case_json = json.load(fp)
-        with open('tests/test_data/collection_exercise/collection_exercise.json') as fp:
-            self.collection_exercise_json = json.load(fp)
-        with open('tests/test_data/collection_exercise/collection_exercise_events.json') as fp:
-            self.collection_exercise_events_json = json.load(fp)
-        with open('tests/test_data/collection_instrument/collection_instrument_eq.json') as fp:
-            self.collection_instrument_json = json.load(fp)
-        with open('tests/test_data/sample/sample_attributes.json') as fp:
-            self.sample_attributes_json = json.load(fp)
-        with open('tests/test_data/survey/survey.json') as fp:
-            self.survey_json = json.load(fp)
-
-        self.action_plan_id = self.case_json['actionPlanId']
-        self.case_id = self.case_json['id']
-        self.case_group_id = self.case_json['caseGroup']['id']
-        self.case_ref = self.case_json['caseRef']
-        self.collection_exercise_id = self.collection_exercise_json['id']
-        self.collection_exercise_ref = self.collection_exercise_json['exerciseRef']
-        self.collection_exercise_user_desc = self.collection_exercise_json['userDescription']
-        self.collection_instrument_id = self.collection_instrument_json['id']
-        self.eq_id = self.collection_instrument_json['classifiers']['eq_id']
-        self.form_type = self.collection_instrument_json['classifiers']['form_type']
-        self.jti = str(uuid.uuid4())
-        self.iac_code = ''.join([str(n) for n in range(11)])
-        self.iac1, self.iac2, self.iac3 = self.iac_code[:4], self.iac_code[4:8], self.iac_code[8:]
-        self.iac_json = {'active': '1', 'caseId': self.case_id}
-        self.ru_name = self.sample_attributes_json['attributes']['Prem1']
-        self.sample_unit_id = self.sample_attributes_json['id']
-        self.sample_unit_ref = self.case_json['caseGroup']['sampleUnitRef']
-        self.sample_unit_type = self.case_json['sampleUnitType']
-        self.survey_id = self.survey_json['id']
-        self.survey_ref = self.survey_json['surveyRef']
-        self.eq_payload = {
-            "jti": self.jti,
-            "tx_id": self.jti,
-            "user_id": self.sample_unit_id,
-            "iat": int(time.time()),
-            "exp": int(time.time() + (5 * 60)),
-            "eq_id": self.eq_id,
-            "period_id": self.collection_exercise_ref,
-            "form_type": self.form_type,
-            "collection_exercise_sid": self.collection_exercise_id,
-            "ru_ref": self.sample_unit_ref,
-            "ru_name": self.ru_name,
-            "case_id": self.case_id,
-            "case_ref": self.case_ref,
-            "account_service_url": self.app['ACCOUNT_SERVICE_URL'],
-            "region_code": self.region_code,
-            "language_code": self.language_code,
-            "return_by": self.return_by,
-            "ref_p_end_date": self.end_date,
-            "ref_p_start_date": self.start_date,
-            "display_address": self.sample_attributes_json['attributes']['Prem1'],
-            "prem_1": self.sample_attributes_json['attributes']['Prem1'],
-            "prem_2": self.sample_attributes_json['attributes']['Prem2'],
-            "prem_3": self.sample_attributes_json['attributes']['Prem3'],
-            "prem_4": self.sample_attributes_json['attributes']['Prem4'],
-            "district": self.sample_attributes_json['attributes']['District'],
-            "post_town": self.sample_attributes_json['attributes']['PostTown'],
-            "postcode": self.sample_attributes_json['attributes']['Postcode'],
-            "country_code": self.sample_attributes_json['attributes']['CountryCode'],
-            "reference": self.sample_attributes_json['attributes']['Reference'],
-        }
-
-        self.case_url = (
-            f"{self.app['CASE_URL']}/cases/{self.case_id}"
-        )
-        self.case_events_url = (
-            f"{self.app['CASE_URL']}/cases/{self.case_id}/events"
-        )
-        self.collection_instrument_url = (
-            f"{self.app['COLLECTION_INSTRUMENT_URL']}"
-            f"/collection-instrument-api/1.0.2/collectioninstrument/id/{self.collection_instrument_id}"
-        )
-        self.collection_exercise_url = (
-            f"{self.app['COLLECTION_EXERCISE_URL']}"
-            f"/collectionexercises/{self.collection_exercise_id}"
-        )
-        self.collection_exercise_events_url = (
-            f"{self.app['COLLECTION_EXERCISE_URL']}"
-            f"/collectionexercises/{self.collection_exercise_id}/events"
-        )
-        self.iac_url = (
-            f"{self.app['IAC_URL']}/iacs/{self.iac_code}"
-        )
-        self.sample_attributes_url = (
-            f"{self.app['SAMPLE_URL']}/samples/{self.sample_unit_id}/attributes"
-        )
-        self.survey_url = (
-            f"{self.app['SURVEY_URL']}/surveys/{self.survey_id}"
-        )
-
-        self.form_data = {
-            'iac1': self.iac1, 'iac2': self.iac2, 'iac3': self.iac3, 'action[save_continue]': '',
-        }
-
-    @unittest_run_loop
-    async def test_get_index(self):
-        response = await self.client.request("GET", "/")
-        self.assertEqual(response.status, 200)
-        contents = await response.content.read()
-        self.assertIn(b'Enter your unique access code below', contents)
-        self.assertEqual(contents.count(b'input-text'), 3)
-        self.assertIn(b'type="submit"', contents)
-
-    @skip_build_eq
-    @unittest_run_loop
-    async def test_post_index(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.post(self.case_events_url)
-
-            with self.assertLogs('respondent-home', 'INFO') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, 'Redirecting to eQ')
-
-        self.assertEqual(response.status, 302)
-        self.assertIn(self.app['EQ_URL'], response.headers['location'])
-
-    @unittest_run_loop
-    async def test_post_index_connector_error(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, exception=ClientConnectorError(mock.MagicMock(), mock.MagicMock()))
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Service connection error")
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(CONNECTION_ERROR_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_no_iac_json(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, content_type='text')
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Service failed to return expected JSON payload")
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(SERVER_ERROR_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_case_403(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, status=403)
-
-            with self.assertLogs('app.case', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Error retrieving case", case_id=self.case_id, status_code=403)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'403 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_case_500(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, status=500)
-
-            with self.assertLogs('app.case', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Error retrieving case", case_id=self.case_id, status_code=500)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'500 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_case_connector_error(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, exception=ClientConnectorError(mock.MagicMock(), mock.MagicMock()))
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Service connection error")
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(CONNECTION_ERROR_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_with_build_connection_error(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.get(self.collection_instrument_url, exception=ClientConnectionError('Failed'))
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Service connection error", message='Failed')
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(CONNECTION_ERROR_MSG.encode(), await response.content.read())
-
-    @skip_encrypt
-    @unittest_run_loop
-    async def test_post_index_with_build(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            # mocks for initial data setup in post
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.post(self.case_events_url)
-            # mocks for the payload builder
-            mocked.get(self.collection_instrument_url, payload=self.collection_instrument_json)
-            mocked.get(self.collection_exercise_url, payload=self.collection_exercise_json)
-            mocked.get(self.collection_exercise_events_url, payload=self.collection_exercise_events_json)
-            mocked.get(self.sample_attributes_url, payload=self.sample_attributes_json)
-            mocked.get(self.survey_url, payload=self.survey_json)
-
-            with self.assertLogs('respondent-home', 'INFO') as logs_home, self.assertLogs('app.eq', 'INFO') as logs_eq:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(logs_home, 'Redirecting to eQ')
-
-        self.assertEqual(response.status, 302)
-        redirected_url = response.headers['location']
-        self.assertTrue(redirected_url.startswith(self.app['EQ_URL']), redirected_url)  # outputs url on fail
-        _, _, _, query, *_ = urlsplit(redirected_url)  # we only care about the query string
-        token = json.loads(parse_qs(query)['token'][0])  # convert token to dict
-        self.assertLogLine(logs_eq, '', payload=token)  # make sure the payload is logged somewhere
-        self.assertEqual(self.eq_payload.keys(), token.keys())  # fail early if payload keys differ
-        for key in self.eq_payload.keys():
-            if key in ['jti', 'tx_id', 'iat', 'exp']:
-                continue  # skip uuid / time generated values
-            self.assertEqual(self.eq_payload[key], token[key], key)  # outputs failed key as msg
-
-    @build_eq_raises
-    @unittest_run_loop
-    async def test_post_index_build_raises_InvalidEqPayLoad(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            # mocks for initial data setup in post
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.post(self.case_events_url)
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                # decorator makes URL constructor raise InvalidEqPayLoad when build() is called in handler
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Service failed to build eQ payload")
-
-        # then error handler catches exception and flashes message to index
-        self.assertEqual(response.status, 200)
-        self.assertIn(REDIRECT_FAILED_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_with_build_ci_500(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            # mocks for initial data setup in post
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.post(self.case_events_url)
-            # mocks for the payload builder
-            mocked.get(self.collection_instrument_url, status=500)
-
-            with self.assertLogs('app.eq', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Error in response", status_code=500)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'500 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_with_build_ci_400(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            # mocks for initial data setup in post
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.post(self.case_events_url)
-            # mocks for the payload builder
-            mocked.get(self.collection_instrument_url, status=400)
-
-            with self.assertLogs('app.eq', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Error in response", status_code=400)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'400 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_with_build_ce_503(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            # mocks for initial data setup in post
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.post(self.case_events_url)
-            # mocks for the payload builder
-            mocked.get(self.collection_instrument_url, payload=self.collection_instrument_json)
-            mocked.get(self.collection_exercise_url, status=503)
-
-            with self.assertLogs('app.eq', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Error in response", status_code=503)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'503 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_with_build_events_404(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            # mocks for initial data setup in post
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.post(self.case_events_url)
-            # mocks for the payload builder
-            mocked.get(self.collection_instrument_url, payload=self.collection_instrument_json)
-            mocked.get(self.collection_exercise_url, payload=self.collection_exercise_json)
-            mocked.get(self.collection_exercise_events_url, status=404)
-
-            with self.assertLogs('app.eq', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Error in response", status_code=404)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'404 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_with_build_sample_403(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            # mocks for initial data setup in post
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            mocked.post(self.case_events_url)
-            # mocks for the payload builder
-            mocked.get(self.collection_instrument_url, payload=self.collection_instrument_json)
-            mocked.get(self.collection_exercise_url, payload=self.collection_exercise_json)
-            mocked.get(self.collection_exercise_events_url, payload=self.collection_exercise_events_json)
-            mocked.get(self.sample_attributes_url, status=403)
-
-            with self.assertLogs('app.eq', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Error in response", status_code=403)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'403 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_with_build_case_event_500(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            # mocks for initial data setup in post
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=self.case_json)
-            # mocks for the payload builder
-            mocked.get(self.collection_instrument_url, payload=self.collection_instrument_json)
-            mocked.get(self.collection_exercise_url, payload=self.collection_exercise_json)
-            mocked.get(self.collection_exercise_events_url, payload=self.collection_exercise_events_json)
-            mocked.get(self.sample_attributes_url, payload=self.sample_attributes_json)
-            mocked.get(self.survey_url, payload=self.survey_json)
-            mocked.post(self.case_events_url, status=500)
-
-            with self.assertLogs('app.case', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, "Error posting case event", status_code=500, case_id=self.case_id)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'500 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_caseid_missing(self):
-        iac_json = self.iac_json.copy()
-        del iac_json['caseId']
-
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=iac_json)
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, 'caseId missing from IAC response')
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(BAD_RESPONSE_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_sampleUnitType_missing(self):
-        case_json = self.case_json.copy()
-        del case_json['sampleUnitType']
-
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=case_json)
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, 'sampleUnitType missing from case response')
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(BAD_RESPONSE_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_sampleUnitType_B_error(self):
-        case_json = self.case_json.copy()
-        case_json['sampleUnitType'] = 'B'
-
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=self.iac_json)
-            mocked.get(self.case_url, payload=case_json)
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", allow_redirects=False, data=self.form_data)
-            self.assertLogLine(cm, 'Attempt to use unexpected sample unit type', sample_unit_type='B')
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(INVALID_CODE_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_malformed(self):
-        form_data = self.form_data.copy()
-        del form_data['iac3']
-
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=self.iac_json)
-
-            with self.assertLogs('respondent-home', 'WARNING') as cm:
-                response = await self.client.request("POST", "/", data=form_data)
-            self.assertLogLine(cm, "Attempt to use a malformed access code")
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(BAD_CODE_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_active_missing(self):
-        iac_json = self.iac_json.copy()
-        del iac_json['active']
-
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=iac_json)
-
-            with self.assertLogs('respondent-home', 'INFO') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Attempt to use an inactive access code")
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(CODE_USED_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_inactive(self):
-        iac_json = self.iac_json.copy()
-        iac_json['active'] = False
-
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, payload=iac_json)
-
-            with self.assertLogs('respondent-home', 'INFO') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Attempt to use an inactive access code")
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(CODE_USED_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_service_connection_error(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, exception=ClientConnectionError('Failed'))
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Client failed to connect to iac service")
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(CONNECTION_ERROR_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_service_500(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, status=500)
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Error in response", status_code=500)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'500 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_service_503(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, status=503)
-
-            with self.assertLogs('respondent-home', 'ERROR') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Error in response", status_code=503)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(f'503 {SERVER_ERROR_MSG}'.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_service_404(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, status=404)
-
-            with self.assertLogs('respondent-home', 'INFO') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Attempt to use an invalid access code", client_ip=None)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(BAD_CODE_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_service_403(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, status=403)
-
-            with self.assertLogs('respondent-home', 'INFO') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Unauthorized access to IAC service attempted", client_ip=None)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(NOT_AUTHORIZED_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_service_401(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, status=401)
-
-            with self.assertLogs('respondent-home', 'INFO') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Unauthorized access to IAC service attempted", client_ip=None)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(NOT_AUTHORIZED_MSG.encode(), await response.content.read())
-
-    @unittest_run_loop
-    async def test_post_index_iac_service_400(self):
-        with aioresponses(passthrough=[str(self.server._root)]) as mocked:
-            mocked.get(self.iac_url, status=400)
-
-            with self.assertLogs('respondent-home', 'INFO') as cm:
-                response = await self.client.request("POST", "/", data=self.form_data)
-            self.assertLogLine(cm, "Client error when accessing IAC service", client_ip=None, status=400)
-
-        self.assertEqual(response.status, 200)
-        self.assertIn(BAD_RESPONSE_MSG.encode(), await response.content.read())
-
-    def test_join_iac(self):
-        # Given some post data
-        post_data = {'iac1': '1234', 'iac2': '5678', 'iac3': '9012', 'action[save_continue]': ''}
-
-        # When join_iac is called
-        result = join_iac(post_data)
-
-        # Then a single string built from the iac values is returned
-        self.assertEqual(result, post_data['iac1'] + post_data['iac2'] + post_data['iac3'])
-
-    def test_join_iac_missing(self):
-        # Given some missing post data
-        post_data = {'action[save_continue]': ''}
-
-        # When join_iac is called
-        with self.assertRaises(TypeError):
-            join_iac(post_data)
-        # Then a TypeError is raised
-
-    def test_join_iac_some_missing(self):
-        # Given some missing post data
-        post_data = {'iac1': '1234', 'iac2': '5678', 'iac3': '', 'action[save_continue]': ''}
-
-        # When join_iac is called
-        with self.assertRaises(TypeError):
-            join_iac(post_data)
-        # Then a TypeError is raised
-
-    def test_correct_iso8601_date_format(self):
-        # Given a valid date
-        date = '2007-01-25T12:00:00Z'
-
-        # When format_date is called
-        result = format_date(date)
-
-        # Then the date is formatted correctly
-        self.assertEqual(result, '2007-01-25')
-
-    def test_incorrect_date_format(self):
-        # Given an invalid date
-        date = 'invalid_date'
-
-        # When format_date is called
-        with self.assertRaises(InvalidEqPayLoad) as e:
-            format_date(date)
-
-        # Then an InvalidEqPayLoad is raised
-        self.assertEqual(e.exception.message, 'Unable to format invalid_date')
-
-    def test_validate_case(self):
-        # Given a dict with an active key and value
-        case_json = {'active': True}
-
-        # When validate_case is called
-        validate_case(case_json)
-
-        # Nothing happens
-
-    def test_validate_case_inactive(self):
-        # Given a dict with an active key and value
-        case_json = {'active': False}
-
-        # When validate_case is called
-        with self.assertRaises(InactiveCaseError):
-            validate_case(case_json)
-
-        # Then an InactiveCaseError is raised
-
-    def test_validate_case_empty(self):
-        # Given an empty dict
-        case_json = {}
-
-        # When validate_case is called
-        with self.assertRaises(InactiveCaseError):
-            validate_case(case_json)
-
-        # Then an InactiveCaseError is raised
+class TestEq(RHTestCase):
 
     def test_create_eq_constructor(self):
         from app import eq
@@ -1172,3 +381,24 @@ class TestGenerateEqURL(AioHTTPTestCase):
              }, "A Second House On The Third Hill"),
         ]:
             self.assertEqual(eq.EqPayloadConstructor.build_display_address(attributes), expected)
+
+    def test_correct_iso8601_date_format(self):
+        # Given a valid date
+        date = '2007-01-25T12:00:00Z'
+
+        # When format_date is called
+        result = format_date(date)
+
+        # Then the date is formatted correctly
+        self.assertEqual(result, '2007-01-25')
+
+    def test_incorrect_date_format(self):
+        # Given an invalid date
+        date = 'invalid_date'
+
+        # When format_date is called
+        with self.assertRaises(InvalidEqPayLoad) as e:
+            format_date(date)
+
+        # Then an InvalidEqPayLoad is raised
+        self.assertEqual(e.exception.message, 'Unable to format invalid_date')
